refactor: Make runner::error private
